extracted/requirements.txt
# MultiClaw Python dependencies

# Web framework + real-time
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0
python-multipart>=0.0.9

# AI model clients (all OpenAI-compatible)
openai>=1.40.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Environment
python-dotenv>=1.0.0

# Utilities
aiofiles>=23.0.0
